+++ src/app/api/indexRequests.ts
@@ -1,0 +1,7 @@
+import * as UserRequest from "./resources/users";
+import * as PostsRequest from "./resources/posts";
+
+export default {
+  UserRequest,
+  PostsRequest,
+};
