+++ src/app/api/resources/posts.ts
@@ -1,0 +1,54 @@
+import axios from "axios";
+
+axios.defaults.baseURL = "https://starforum.herokuapp.com/api/v1";
+
+export const addNewPost = (title: string, body: string, parent: string) => {
+  const datas = {
+    title: title,
+    body: body,
+    parent: parent,
+  };
+  axios.post("/posts", datas).then(response => {
+    console.log(response);
+    return response;
+  });
+};
+
+export const getPosts = async () => {
+  const response = await axios.get("/posts");
+  console.log(response.data);
+  return response.data;
+};
+
+export const getPostByID = async (id: string) => {
+  const response = await axios.get(`/posts/${id}`);
+  console.log(response.data);
+  return response.data;
+};
+
+export const getPostsNumber = async () => {
+  const response = await axios.get("/posts-number");
+  console.log(response.data);
+  return response.data;
+};
+
+export const upvotePost = (id: string) => {
+  axios.post(`/posts/${id}/upvote`).then(response => {
+    console.log(response);
+    return response;
+  });
+};
+
+export const downvotePost = (id: string) => {
+  axios.post(`/posts/${id}/downvote`).then(response => {
+    console.log(response);
+    return response;
+  });
+};
+
+export const removeReaction = (id: string) => {
+  axios.post(`/posts/${id}/remove-reaction`).then(response => {
+    console.log(response);
+    return response;
+  });
+};
