+++ src/app/api/resources/users.ts
@@ -1,6 +1,6 @@
 import axios from "axios";
 
-axios.defaults.baseURL='https://starforum.herokuapp.com/api/v1'
+axios.defaults.baseURL = "https://starforum.herokuapp.com/api/v1";
 
 export const addUser = (name: string, pass: string) => {
   const data = {
@@ -8,23 +8,23 @@ export const addUser = (name: string, pass: string) => {
     password: pass,
   };
 
- axios.post("/users", data);
+  axios.post("/users", data);
 };
 
 export const loginUser = (name: string, pass: string) => {
   const data = {
     username: name,
     password: pass,
-  }
-  fetch('https://starforum.herokuapp.com/api/v1/login', {
-    method: 'POST',
-    credentials: 'include',
+  };
+  fetch("https://starforum.herokuapp.com/api/v1/login", {
+    method: "POST",
+    credentials: "include",
     headers: {
-      'Content-Type': 'application/json',
+      "Content-Type": "application/json",
     },
     body: JSON.stringify(data),
-  })
-}
+  });
+};
 
 export const getToken = async (username: string) => {
   const config: any = {
@@ -33,25 +33,20 @@ export const getToken = async (username: string) => {
       withCredentials: true,
     },
   };
-  const response = await axios.get(
-    `/token/${username}`,
-    config,
-  );
+  const response = await axios.get(`/token/${username}`, config);
   console.log(response);
   return response;
 };
 
 export const logoutUser = () => {
-  axios.delete("/logout").then(response =>{
-    console.log(response)
-    return response
-  })
+  axios.delete("/logout").then(response => {
+    console.log(response);
+    return response;
+  });
 };
 
 export const getUserByUsernameReaction = async (username: string) => {
-  const response = await axios.get(
-    `/users/${username}/reactions`,
-  );
+  const response = await axios.get(`/users/${username}/reactions`);
   console.log(response.data);
   return response.data;
 };
@@ -67,5 +62,5 @@ export const addUserIcon = () => {};
 export const getUserIcon = async (username: string) => {
   const response = await axios.get(`/users/${username}/icon`);
   console.log(response);
-  return response.data;// img in Base64 format
+  return response.data; // img in Base64 format
 };
