--- src/app/service/indexRequests.ts
@@ -1,7 +1,0 @@
-import * as UserRequest from "./users";
-import * as PostsRequest from "./posts";
-
-export default {
-  UserRequest,
-  PostsRequest,
-};
